# manual/static/openapi/admin-api-openapi.json
{
  "openapi": "3.0.2",
...
  "info": {
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "MAIF open source team",
      "email": "oss@maif.fr"
    },
    "description": "Admin API of Daikoku",
    "title": "Daikoku Admin API",
    "version": "16.3.0"
  },
...
  "components": {
    "schemas": {
...
      "done": {
        "description": "task is done",
        "properties": {
          "done": {
            "type": "boolean"
          }
        }
      },
      "error": {
        "description": "error response",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "patch": {
        "description": "A set of changes described in JSON Patch format: http://jsonpatch.com/ (RFC 6902)",
        "type": "array",
        "items": {
          "type": "object",
          "required": [
            "op",
            "path"
          ],
          "properties": {
            "op": {
              "type": "string",
              "enum": [
                "add",
                "replace",
                "remove",
                "copy",
                "test"
              ]
            },
            "path": {
              "type": "string"
            },
            "value": {}
          }
        }
      },
...
      "User": {
        "description": "User",
        "properties": {
          "_id": {
            "type": "string"
          },
          "_deleted": {
            "type": "boolean"
          },
          "tenants": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "origins": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string",
              "enum": [
                "local",
                "otoroshi",
                "ldap",
                "oauth2"
              ]
            }
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "pictureFromProvider": {
            "type": "boolean",
            "nullable": true
          },
          "personalToken": {
            "type": "string",
            "nullable": true
          },
          "isDaikokuAdmin": {
            "type": "boolean",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true,
            "description": "optionnal value but required in case of tenant has a Local AuthProvider"
          },
          "hardwareKeyRegistrations": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object"
            }
          },
          "lastTenant": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "defaultLanguage": {
            "type": "string",
            "nullable": true
          },
          "starredApis": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "twoFactorAuthentication": {
            "$ref": "#/components/schemas/TwoFactorAuthentication",
            "nullable": true
          },
          "invitation": {
            "$ref": "#/components/schemas/UserInvitation",
            "nullable": true
          }
        },
        "required": [
          "_id",
          "name",
          "email"
        ]
      },
      "UserInvitation": {
        "description": "UserInvitation",
        "properties": {
          "token": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "format": "timestamp",
            "nullable": true
          },
          "team": {
            "type": "string"
          },
          "notificationId": {
            "type": "string",
            "nullable": true
          },
          "registered": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "token"
        ]
      },
      "TwoFactorAuthentication": {
        "description": "TwoFactorAuthentication",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "secret": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "backupCodes": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "secret",
          "token",
          "backupCodes"
        ]
      },
...
    }
  },
  "paths": {
...
    "/admin-api/users/{id}": {
      "delete": {
        "summary": "delete a user",
        "operationId": "users.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "patch": {
        "summary": "update a user with JSON patch or by  merging JSON object",
        "operationId": "users.patch",
        "requestBody": {
          "description": "the patch to update the user or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/User"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "put": {
        "summary": "update a user",
        "operationId": "users.update",
        "requestBody": {
          "description": "the user to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "get": {
        "summary": "read a user",
        "operationId": "users.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
    "/admin-api/users": {
      "get": {
        "summary": "read all user",
        "operationId": "users.findAll",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "post": {
        "summary": "creates a user",
        "requestBody": {
          "description": "the user to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "operationId": "users.create",
        "responses": {
          "201": {
            "description": "entity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "bad entity format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
  }
}